add multiline rows
Now one cell of the table can have a multiline string as content. The solution
is not so beautiful, but it works. Also updated the tests.
Beside this I added a separation line between all rows of the table, to a visual
difference between between rows and rows with more than one line as content.